+++ NEWS.md
@@ -1,3 +1,18 @@
+# 0.2.0.9000 - F mayor
+
+## New features
+
+- New website at https://eliocamp.github.io/ggnewscale/
+
+## Bugfixes
+
+* `new_scale()` does not fail when `guide = "none"` or `guide = FALSE` (#6, thanks @Maschette)
+
+* `new_scale()` does not fail when using the default mapping (#7, thanks @
+RomanAbashin)
+
+* `new_scale()` adds a default scale if not supplied.
+
 # ggnewscale 0.2.0 - C major
 
 ## Bugfixes